Augmented e2e test scenarios in one single test suite.
